# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bookd API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Slot already booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bookings/my:
    get:
      operationId: listMyBookings
      tags: [bookings]
      summary: Get current user bookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: size
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
  /bookings/{id}/cancel:
    put:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/complete:
    put:
      operationId: completeBooking
      tags: [bookings]
      summary: Mark booking as complete (PROVIDER only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review for a completed booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    get:
      operationId: listServiceReviews
      tags: [reviews]
      summary: Get reviews for a service (pass serviceId as query param)
      parameters:
        - name: serviceId
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: size
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
...
  /provider/bookings:
    get:
      operationId: listProviderBookings
      tags: [provider]
      summary: Get bookings for provider services
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: size
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Provider bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [CUSTOMER, PROVIDER]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
...
    Service:
      type: object
      properties:
        id:
          type: string
        providerId:
          type: string
        provider:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [HEALTH, BEAUTY, FITNESS, EDUCATION, HOME, OTHER]
        price:
          type: number
        durationMinutes:
          type: integer
        location:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - providerId
        - provider
        - title
        - description
        - category
        - price
        - durationMinutes
        - location
        - isActive
        - reviewCount
        - createdAt
...
    Availability:
      type: object
      properties:
        id:
          type: string
        serviceListingId:
          type: string
        slotDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        isBooked:
          type: boolean
      required:
        - id
        - serviceListingId
        - slotDate
        - startTime
        - endTime
        - isBooked
...
    Booking:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        customer:
          $ref: "#/components/schemas/User"
        serviceListingId:
          type: string
        service:
          $ref: "#/components/schemas/Service"
        availabilityId:
          type: string
        availability:
          $ref: "#/components/schemas/Availability"
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
        hasReview:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - customerId
        - service
        - availabilityId
        - availability
        - status
        - totalPrice
        - hasReview
        - createdAt
    BookingInput:
      type: object
      properties:
        serviceListingId:
          type: string
        availabilityId:
          type: string
        notes:
          type: string
      required:
        - serviceListingId
        - availabilityId
    BookingListResponse:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        totalPages:
          type: integer
      required:
        - bookings
        - total
        - page
        - size
        - totalPages
    Review:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        customerId:
          type: string
        customer:
          $ref: "#/components/schemas/User"
        serviceListingId:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - bookingId
        - customerId
        - customer
        - serviceListingId
        - rating
        - createdAt
    ReviewInput:
      type: object
      properties:
        bookingId:
          type: string
        rating:
          type: integer
        comment:
          type: string
      required:
        - bookingId
        - rating
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - reviews
        - total
        - page
        - totalPages
...