# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bookd API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Availability } from "./availability";
import type { BookingStatus } from "./bookingStatus";
import type { Service } from "./service";
import type { User } from "./user";

export interface Booking {
  id: string;
  customerId: string;
  customer?: User;
  serviceListingId?: string;
  service: Service;
  availabilityId: string;
  availability: Availability;
  status: BookingStatus;
  totalPrice: number;
  /** @nullable */
  notes?: string | null;
  hasReview: boolean;
  createdAt: string;
}