# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bookd API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceCategory } from "./serviceCategory";
import type { User } from "./user";

export interface Service {
  id: string;
  providerId: string;
  provider: User;
  title: string;
  description: string;
  category: ServiceCategory;
  price: number;
  durationMinutes: number;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  /** @nullable */
  averageRating?: number | null;
  reviewCount: number;
  createdAt: string;
}